Break out opentelemetry-api-util artifact
