+++ api/util/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+  id("otel.java-conventions")
+  id("otel.publish-conventions")
+
+  id("otel.animalsniffer-conventions")
+}
+
+description = "OpenTelemetry API Common"
+otelJava.moduleName.set("io.opentelemetry.api.util")
+
+dependencies {
+}
