+++ context/src/main/java/io/opentelemetry/context/LazyStorage.java
@@ -37,6 +37,7 @@
 
 package io.opentelemetry.context;
 
+import io.opentelemetry.api.util.ComponentLoader;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
