+++ docs/apidiffs/current_vs_latest/opentelemetry-api-util.txt
@@ -1,0 +1,7 @@
+Comparing source compatibility of opentelemetry-api-util-1.52.0-SNAPSHOT.jar against 
++++  NEW INTERFACE: PUBLIC(+) ABSTRACT(+) io.opentelemetry.api.util.ComponentLoader  (not serializable)
+	+++  CLASS FILE FORMAT VERSION: 52.0 <- n.a.
+	+++  NEW SUPERCLASS: java.lang.Object
+	+++  NEW METHOD: PUBLIC(+) STATIC(+) io.opentelemetry.api.util.ComponentLoader forClassLoader(java.lang.ClassLoader)
+	+++  NEW METHOD: PUBLIC(+) ABSTRACT(+) java.lang.Iterable<T> load(java.lang.Class<T>)
+		GENERIC TEMPLATES: +++ T:java.lang.Object
