+++ exporters/common/src/test/java/io/opentelemetry/exporter/internal/compression/CompressorUtilTest.java
@@ -8,7 +8,7 @@ package io.opentelemetry.exporter.internal.compression;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import io.opentelemetry.context.ComponentLoader;
+import io.opentelemetry.api.util.ComponentLoader;
 import java.net.URL;
 import java.net.URLClassLoader;
 import org.junit.jupiter.api.Test;
