+++ settings.gradle.kts
@@ -26,6 +26,7 @@ include(":all")
 include(":api:all")
 include(":api:incubator")
 include(":api:testing-internal")
+include(":api:util")
 include(":bom")
 include(":bom-alpha")
 include(":context")
